requirements.txt
matplotlib==3.4.3
numpy==1.24.2
opencv_python==4.6.0.66
Pillow==8.3.2
pycocotools==2.0.6
PyYAML==6.0
torch==1.8.1+cu111
torchmetrics==0.11.1
torchvision==0.9.1+cu111
tqdm==4.61.2
